modules/firefox: allow overriding wrapperArgs; add schizofox.misc.speechSynthesisSupport
programs.schizofox.extraWrapperArgs will allow the user to pass a freeform attribute set to the wrapper as is, in order to merge additional arguments to the initial wrapper. This is especially useful when wrapFirefox has new options that *do not* have options in Schizofox module yet. Do note that this is extremely experimental and very implicit. Expect breakage, and use with great care.